# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Basic CRUD Laravel API Documentation",
        "description": "API Documentation - Basic CRUD Laravel",
        "contact": {
            "email": "manirujjamanakash@gmail.com"
        },
        "license": {
            "name": "GPL2",
            "url": "https://devsenv.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books/view/all": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "All Books - Publicly Accessible",
                "description": "All Books - Publicly Accessible",
                "operationId": "indexAll",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Books - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Show Books Details",
                "description": "Show Books Details",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Books Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update Books",
                "description": "Update Books",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Books 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 10120
                                    },
                                    "file": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Books"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete Books",
                "description": "Delete Books",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Books"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get Devices List",
                "description": "Get Devices List as Array",
                "operationId": "indexDevices",
                "responses": {
                    "200": {
                        "description": "Get Devices List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Create New Devices",
                "description": "Create New Devices",
                "operationId": "storeDevices",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ip": {
                                        "type": "string",
                                        "example": "Devices 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create New Devices"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/devices/view/all": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "All Devices - Publicly Accessible",
                "description": "All Devices - Publicly Accessible",
                "operationId": "indexAllDevices",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Devices - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/devices/view/search": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "All Devices - Publicly Accessible",
                "description": "All Devices - Publicly Accessible",
                "operationId": "searchDevice",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search, eg; Test",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Test"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Devices - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/devices/{id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Show Devices Details",
                "description": "Show Devices Details",
                "operationId": "showDevices",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Devices Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update Devices",
                "description": "Update Devices",
                "operationId": "updateDevices",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ip": {
                                        "type": "string",
                                        "example": "Devices 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Devices"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Delete Devices",
                "description": "Delete Devices",
                "operationId": "destroyDevices",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Devices"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}